+++ R/visualization_functions.R
@@ -313,10 +313,10 @@ visualize_hsa_KEGG <- function(pw_table, gene_data) {
 }
 
 
-#' Plot the Bubble Chart of Enrichment Results
+#' Create Bubble Chart of Enrichment Results
 #'
-#' This function is used to plot a bubble chart displaying the enrichment
-#' results.
+#' This function is used to create a ggplot2 bubble chart displaying the
+#' enrichment results.
 #'
 #' @param result_df a data frame that must contain the following columns: \describe{
 #'   \item{Term_Description}{Description of the enriched term}
@@ -326,6 +326,11 @@ visualize_hsa_KEGG <- function(pw_table, gene_data) {
 #'   \item{Down_regulated}{the down-regulated genes in the input involved in the given term's gene set, comma-separated}
 #'   \item{Cluster(OPTIONAL)}{the cluster to which the enriched term is assigned}
 #' }
+#' @param top_terms number of top terms (according to the "lowest_p" column)
+#'  to plot (default = 10). If \code{plot_by_cluster = TRUE}, selects the top
+#'  \code{top_terms} terms per each cluster. Set \code{top_terms = NULL} to plot
+#'  for all terms.If the total number of terms is less than \code{top_terms},
+#'  all terms are plotted.
 #' @param plot_by_cluster boolean value indicating whether or not to group the
 #'  enriched terms by cluster (works if \code{result_df} contains a
 #'  "Cluster" column).
@@ -350,67 +355,79 @@ visualize_hsa_KEGG <- function(pw_table, gene_data) {
 #'
 #' @examples
 #' g <- enrichment_chart(RA_output)
-enrichment_chart <- function(result_df, plot_by_cluster = FALSE,
-                             num_bubbles = 4, even_breaks = TRUE) {
-  necessary <- c(
-    "Term_Description", "Fold_Enrichment", "lowest_p",
-    "Up_regulated", "Down_regulated"
-  )
+enrichment_chart <- function(result_df,
+                             top_terms = 10,
+                             plot_by_cluster = FALSE,
+                             num_bubbles = 4,
+                             even_breaks = TRUE) {
+  necessary <- c("Term_Description", "Fold_Enrichment", "lowest_p",
+                 "Up_regulated", "Down_regulated")
+
   if (!all(necessary %in% colnames(result_df))) {
-    stop(
-      "The input data frame must have the columns:\n",
-      paste(necessary, collapse = ", ")
-    )
+    stop("The input data frame must have the columns:\n",
+         paste(necessary, collapse = ", "))
   }
 
   if (!is.logical(plot_by_cluster)) {
     stop("`plot_by_cluster` must be either TRUE or FALSE")
   }
 
+  if (!is.numeric(top_terms) & !is.null(top_terms)) {
+    stop("`top_terms` must be either numeric or NULL")
+  }
+
+  if (!is.null(top_terms)){
+    if (top_terms < 1) {
+      stop("`top_terms` must be > 1")
+    }
+  }
+
   # sort by lowest adj.p
   result_df <- result_df[order(result_df$lowest_p), ]
 
-  num_genes <- vapply(
-    result_df$Up_regulated,
-    function(x) length(unlist(strsplit(x, ", "))), 1
-  )
-  num_genes <- num_genes + vapply(
-    result_df$Down_regulated,
-    function(x) length(unlist(strsplit(x, ", "))), 1
-  )
+  ## Filter for top_terms
+  if (!is.null(top_terms)) {
+    if (plot_by_cluster) {
+      keep_ids <- tapply(result_df$ID, result_df$Cluster, function(x)
+                                          x[seq_len(min(top_terms, length(x)))])
+      keep_ids <- unlist(keep_ids)
+      result_df <- result_df[result_df$ID %in% keep_ids, ]
+    } else if (top_terms < nrow(result_df)){
+      result_df <- result_df[seq_len(top_terms), ]
+    }
+  }
+
+  num_genes <- vapply(result_df$Up_regulated,
+    function(x) length(unlist(strsplit(x, ", "))), 1)
+  num_genes <- num_genes + vapply(result_df$Down_regulated,
+    function(x) length(unlist(strsplit(x, ", "))), 1)
 
   result_df$Term_Description <- factor(result_df$Term_Description,
-    levels = rev(unique(result_df$Term_Description))
-  )
+          levels = rev(unique(result_df$Term_Description)))
 
   g <- ggplot2::ggplot(result_df, ggplot2::aes_(~Fold_Enrichment, ~Term_Description))
-  g <- g + ggplot2::geom_point(ggplot2::aes(
-    color = -log10(result_df$lowest_p),
-    size = num_genes
-  ), na.rm = TRUE)
+  g <- g + ggplot2::geom_point(ggplot2::aes(color = -log10(result_df$lowest_p),
+                                            size = num_genes), na.rm = TRUE)
   g <- g + ggplot2::theme_bw()
-  g <- g + ggplot2::theme(
-    axis.text.x = ggplot2::element_text(size = 10),
-    axis.text.y = ggplot2::element_text(size = 10),
-    plot.title = ggplot2::element_blank()
-  )
+  g <- g + ggplot2::theme(axis.text.x = ggplot2::element_text(size = 10),
+                          axis.text.y = ggplot2::element_text(size = 10),
+                          plot.title = ggplot2::element_blank())
   g <- g + ggplot2::xlab("Fold Enrichment")
   g <- g + ggplot2::theme(axis.title.y = ggplot2::element_blank())
-  g <- g + ggplot2::labs(size = "# of DEGs", color = "-log10(lowest-p)")
+  g <- g + ggplot2::labs(size = "# of DEGs", color = expression(-log[10](p)))
 
   ## breaks for # of DEGs
   if (max(num_genes) < num_bubbles) {
     g <- g + ggplot2::scale_size_continuous(breaks = seq(0, max(num_genes)))
   } else {
-    tmp1 <- base::seq(
-      0, max(num_genes),
-      round(max(num_genes) / (num_bubbles + 1))
-    )
-    tmp2 <- base::round(base::seq(0, max(num_genes),
-      length.out = num_bubbles + 1
-    ))
-    brks <- ifelse(even_breaks, tmp1, tmp2)
 
+    if (even_breaks) {
+      brks <- base::seq(0, max(num_genes),
+                        round(max(num_genes) / (num_bubbles + 1)))
+    } else {
+      brks <- base::round(base::seq(0, max(num_genes),
+                                    length.out = num_bubbles + 1))
+    }
     g <- g + ggplot2::scale_size_continuous(breaks = brks)
   }